Schedule_Optimizer/scraper/requirements.txt
certifi==2020.6.20
chardet==3.0.4
flake8==3.8.3
idna==2.10
lxml==4.5.2
mccabe==0.6.1
mysql-connector-python==8.0.21
protobuf==3.13.0
pycodestyle==2.6.0
pyflakes==2.2.0
requests==2.24.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
